Bugfixes:
 - Trace unit now outputs reasonably sane trace information when the core is
   running with stop bit enabled code.

Known issues/TODOs:

 - For some reason the valid trace data is followed by a couple kb of garbage
   when the core is executing stop bit enabled code. No idea why this is
   happening yet... seems to be fine with the same bit file when running code
   assembled with --padding=8.

+++ tools/debug-interface/src/rvtrace/traceParse.h
@@ -110,7 +110,7 @@ typedef struct {
    * Reconfiguration information.
    */
   int hasNewConfiguration;
-  uint32_t newConfiguration;
+  uint8_t newConfiguration;
   
 } trace_packet_t;
